Moved sqlite specific code into sqlite module
