+++ discovery_engine_core/core/src/storage.rs
@@ -12,8 +12,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use std::borrow::Cow;
-
 use async_trait::async_trait;
 use displaydoc::Display;
 use thiserror::Error;
@@ -38,22 +36,6 @@ pub enum Error {
 impl From<sqlx::Error> for Error {
     fn from(generic: sqlx::Error) -> Self {
         Error::Database(generic.into())
-    }
-}
-
-trait SqlxSqliteErrorExt<V> {
-    fn fk_violation_is_invalid_document_id(self, id: document::Id) -> Result<V, Error>;
-}
-
-impl<V> SqlxSqliteErrorExt<V> for Result<V, sqlx::Error> {
-    fn fk_violation_is_invalid_document_id(self, id: document::Id) -> Result<V, Error> {
-        if let Err(sqlx::Error::Database(db_err)) = &self {
-            if db_err.code() == Some(Cow::Borrowed("787")) {
-                return Err(Error::InvalidDocumentId(id));
-            }
-        }
-        self.map_err(Into::into)
-
     }
 }
